pyproject.toml
[tool.poetry]
name = "scrapping"
version = "0.0.1"
description = "none"
authors = ["Your Name <you@example.com>"]
license = "none"

[tool.poetry.dependencies]
python = "<3.12,>=3.8"
numpy = "^1.23.4"
fake-useragent = "^0.1.13"
pandas = "^1.5.1"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
mysql-connector-python = "^8.0.31"
logging = "^0.4.9.6"
lxml = "^4.9.1"
elementpath = "^3.0.2"
scipy = "^1.10.0"
matplotlib = "^3.6.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
